+++ src/utils/index.js
@@ -1,5 +1,6 @@
-const {createRequiredApprovalsComment} = require('./create-required-approval-comment');
 const {createOwnersMap} = require('./create-owners-map');
+const {createRequiredApprovalsComment} = require('./create-required-approval-comment');
+const {createUpdatedBody} = require('./create-updated-body');
 const {filterChangedFiles} = require('./filter-changed-files');
 const {getListReviewers} = require('./get-list-reviewers');
 const {getLatestUserReviewMap} = require('./get-latest-user-review-map');
@@ -9,8 +10,9 @@ const {getOwners} = require('./get-owners');
 const {validateLabelsMap} = require('./validate-labels-map');
 
 module.exports = {
-    createRequiredApprovalsComment,
     createOwnersMap,
+    createRequiredApprovalsComment,
+    createUpdatedBody,
     filterChangedFiles,
     getListReviewers,
     getLatestUserReviewMap,