+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- [#36](https://github.com/zattoo/reviewers/issues/36) Show required reviewers on Pull Request description
+
 ## 3.0.1 - 220.12.2021
 
 ### Fixed